feat: add LoginLoader component and sanitizeInput utility function
