+++ src/app/login/components/LoginLoader.tsx
@@ -1,0 +1,17 @@
+function LoginLoader() {
+  return (
+    <div className="min-h-full bg-gray-50 dark:bg-gray-900 flex items-center justify-center py-12 px-4 sm:px-6 lg:px-8">
+      <div className="max-w-md w-full space-y-8 text-center">
+        <div className="animate-pulse">
+          <div className="h-8 bg-gray-200 dark:bg-gray-700 rounded my-4 mx-auto w-3/4" />
+          <div className="h-4 bg-gray-200 dark:bg-gray-700 rounded my-2 mx-auto w-2/4" />
+          <div className="h-12 bg-gray-200 dark:bg-gray-700 rounded-md my-6" />
+          <div className="h-12 bg-gray-200 dark:bg-gray-700 rounded-md my-6" />
+          <div className="h-12 bg-gray-200 dark:bg-gray-700 rounded-md my-6" />
+        </div>
+      </div>
+    </div>
+  )
+}
+
+export default LoginLoader
