+++ src/app/login/utils/sanitizeInput.ts
@@ -1,0 +1,5 @@
+function sanitizeInput(input: string): string {
+  return input.trim().replace(/[<>]/g, '')
+}
+
+export default sanitizeInput
